% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgref-package.R
\docType{package}
\name{avgref-package}
\alias{avgref}
\alias{avgref-package}
\title{avgref: Average Calculation of Soil Spectral Reflectance Data}
\description{
The package calculates the average value of each soil sample in a data set containing the spectral reflectance of post-dried soil samples collected from the Hudson Bay Lowlands. This calculates the reflectance of soil samples at a given wavelength range, which allows specific spectral regions to be compared with other soil metrics.

This package is an exercise in package creation using
R studio. The package includes a dataset of the reflectance values of soil
samples that were collected from the Hudson Bay Lowlands on July 2023. The
package also includes a function for calculating the average reflectance of
each soil sample at a given wavelength range.
}
\references{
\url{https://github.com/KangyuSo/my-GEO712-repository}
}
\author{
\strong{Maintainer}: Kangyu So \email{sok11@mcmaster.ca}


Kangyu So, School of Earth, Environment and Society, McMaster University \email{sok11@mcmaster.ca}
}
\keyword{internal}
