% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgref.R
\name{avgref}
\alias{avgref}
\title{Calculate the average reflectance of soil samples at a given wavelength range}
\source{
The data was collected by Kangyu So, Dr. Alemu Gonsamo, and Dr. Cheryl A. Rogers (All affiliated with McMaster University) from the Hudson Bay Lowlands region on July 2023
}
\usage{
data(ref_dry)
}
\arguments{
\item{start_row}{The lower end of the wavelength range}

\item{end_row}{The higher end of the wavelength range}
}
\value{
The average reflectance of each soil sample at the wavelength range
}
\description{
Calculate the average reflectance of soil samples at a given wavelength range
}
\examples{
data(ref_dry)
ref_dry_average <- avgref(ref_dry, 500, 750)
}
