% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgref-package.R
\docType{data}
\name{ref_dry}
\alias{ref_dry}
\title{Reflectance values for post-dried soil samples.}
\format{
A data frame with 2151 rows and 433 variables:
\describe{
\item{Sample_ID}{ID of the soil sample}
\item{Reflectance}{Spectral reflectance value at a given wavelength}
}
}
\usage{
data(ref_dry)
}
\description{
A dataset containing information about the spectral reflectance of 432
post-dried soil samples collected from the Hudson Bay Lowlands. The range of
information is between the wavelengths 350 nm and 2500 nm in increments of
1 nm.
}
\examples{
 data(ref_dry)
 Wavelength <- ref_dry$Wavelength
}
\keyword{Soil}
\keyword{carbon}
\keyword{datasets}
\keyword{spectroscopy}
\keyword{stock}
